paginacion, autenticacion en consultas y filtrado de clientes

+++ servidor.js
@@ -72,7 +72,7 @@ const verificar_jwt = express.Router();
 verificar_usuario.use((request, response, next) => {
   let usuario = {};
   usuario.nombre = request.body.nombreUsuario;
-  usuario.clave = request.body.clave; 
+  usuario.clave = request.body.clave;
   request.getConnection((err, conn) => {
     if (err) throw "Error al conectarse a la base de datos.";
     conn.query(
@@ -126,7 +126,6 @@ app.post("/login", verificar_usuario, (request, response, obj) => {
 
 verificar_jwt.use((request, response, next) => {
   //SE RECUPERA EL TOKEN DEL ENCABEZADO DE LA PETICIÓN
-  console.log(request.body.headers["authorization"]);
   let token = request.headers["x-access-token"] || request.headers["authorization"];
   if (!token) {
     response.status(401).send({
@@ -191,12 +190,14 @@ app.get("/login", (request, response) => {
 
 // CRUD CLIENTES **************************************************************************************
 // Agregar cliente
-app.post("/agregarCliente", (request, response) => {
+app.post("/agregarCliente", verificar_jwt, (request, response) => {
   let obj_respuesta = {
     exito: false,
     mensaje: "No se pudo agregar el cliente",
     status: 418,
   };
+
+  let jwt = response.jwt;
 
   let cliente_json = {};
   cliente_json.id = request.body.id;
@@ -214,33 +215,39 @@ app.post("/agregarCliente", (request, response) => {
   let control = request.body.control[0];
   control.id = cliente_json.id;
 
-  request.getConnection((err, conn) => {
-    if (err) throw "Error al conectarse a la base de datos.";
-    conn.query("INSERT INTO clientes set ?", [cliente_json], (err, rows) => {
-      if (err) {
-        console.log(err);
-        throw "Error en consulta de base de datos.";
-      }
+  if (jwt.usuario.perfil !== "administrador") {
+    obj_respuesta.mensaje = "Usuario sin permisos!!";
+    obj_respuesta.status = 401;
+    response.status(obj_respuesta.status).json(obj_respuesta);
+  } else {
+    request.getConnection((err, conn) => {
+      if (err) throw "Error al conectarse a la base de datos.";
+      conn.query("INSERT INTO clientes set ?", [cliente_json], (err, rows) => {
+        if (err) {
+          console.log(err);
+          throw "Error en consulta de base de datos.";
+        }
+      });
     });
-  });
 
-  request.getConnection((err, conn) => {
-    if (err) throw "Error al conectarse a la base de datos.";
-    conn.query("INSERT INTO controles set ?", [control], (err, rows) => {
-      if (err) {
-        console.log(err);
-        throw "Error en consulta de base de datos.";
-      }
-      obj_respuesta.exito = true;
-      obj_respuesta.mensaje = "Cliente agregado!";
-      obj_respuesta.status = 200;
-      response.status(obj_respuesta.status).json(obj_respuesta);
+    request.getConnection((err, conn) => {
+      if (err) throw "Error al conectarse a la base de datos.";
+      conn.query("INSERT INTO controles set ?", [control], (err, rows) => {
+        if (err) {
+          console.log(err);
+          throw "Error en consulta de base de datos.";
+        }
+        obj_respuesta.exito = true;
+        obj_respuesta.mensaje = "Cliente agregado!";
+        obj_respuesta.status = 200;
+        response.status(obj_respuesta.status).json(obj_respuesta);
+      });
     });
-  });
+  }
 });
 
 // Listar clientes
-app.get("/listarClientes",(request, response) => {
+app.get("/listarClientes", verificar_jwt, (request, response) => {
   let obj_respuesta = {
     exito: false,
     mensaje: "No se encontraron clientes",
@@ -251,31 +258,39 @@ app.get("/listarClientes",(request, response) => {
 
   let jwt = response.jwt;
 
-  request.getConnection((err, conn) => {
-    if (err) throw "Error al conectarse a la base de datos.";
-    conn.query("SELECT * FROM clientes", (err, rows) => {
-      if (err) throw "Error en consulta de base de datos.";
-      if (rows.length == 0) {
-        response.status(obj_respuesta.status).json(obj_respuesta);
-      } else {
-        obj_respuesta.exito = true;
-        obj_respuesta.mensaje = "Clientes encontrados!";
-        obj_respuesta.dato = rows;
-        obj_respuesta.payload = jwt;
-        obj_respuesta.status = 200;
-        response.status(obj_respuesta.status).json(obj_respuesta);
-      }
+  if (jwt.usuario.perfil !== "administrador") {
+    obj_respuesta.mensaje = "Usuario sin permisos!!";
+    obj_respuesta.status = 401;
+    response.status(obj_respuesta.status).json(obj_respuesta);
+  } else {
+    request.getConnection((err, conn) => {
+      if (err) throw "Error al conectarse a la base de datos.";
+      conn.query("SELECT * FROM clientes", (err, rows) => {
+        if (err) throw "Error en consulta de base de datos.";
+        if (rows.length == 0) {
+          response.status(obj_respuesta.status).json(obj_respuesta);
+        } else {
+          obj_respuesta.exito = true;
+          obj_respuesta.mensaje = "Clientes encontrados!";
+          obj_respuesta.dato = rows;
+          obj_respuesta.payload = jwt;
+          obj_respuesta.status = 200;
+          response.status(obj_respuesta.status).json(obj_respuesta);
+        }
+      });
     });
-  });
+  }
 });
 
 // Modificar cliente
-app.post("/modificarCliente", (request, response) => {
+app.post("/modificarCliente", verificar_jwt, (request, response) => {
   let obj_respuesta = {
     exito: false,
     mensaje: "No se pudo modificar el cliente",
     status: 418,
   };
+
+  let jwt = response.jwt;
 
   let cliente_json = {};
   cliente_json.id = request.body.id;
@@ -292,77 +307,93 @@ app.post("/modificarCliente", (request, response) => {
 
   let control = request.body.control;
 
-  control.forEach((c) => {
+  if (jwt.usuario.perfil !== "administrador") {
+    obj_respuesta.mensaje = "Usuario sin permisos!!";
+    obj_respuesta.status = 401;
+    response.status(obj_respuesta.status).json(obj_respuesta);
+  } else {
+    control.forEach((c) => {
+      request.getConnection((err, conn) => {
+        if (err) throw "Error al conectarse a la base de datos.";
+        conn.query("UPDATE controles SET ? WHERE id = ? AND fecha = ?", [c, c.id, c.fecha], (err, rows) => {
+          if (err) {
+            console.log(err);
+            throw "Error en consulta de base de datos.";
+          }
+        });
+      });
+    });
+
     request.getConnection((err, conn) => {
       if (err) throw "Error al conectarse a la base de datos.";
-      conn.query("UPDATE controles SET ? WHERE id = ? AND fecha = ?", [c, c.id, c.fecha], (err, rows) => {
+      conn.query("UPDATE clientes SET ? WHERE id = ?", [cliente_json, cliente_json.id], (err, rows) => {
         if (err) {
           console.log(err);
           throw "Error en consulta de base de datos.";
         }
+        obj_respuesta.exito = true;
+        obj_respuesta.mensaje = "Cliente Modificado!";
+        obj_respuesta.status = 200;
+        response.status(obj_respuesta.status).json(obj_respuesta);
       });
     });
-  });
-
-  request.getConnection((err, conn) => {
-    if (err) throw "Error al conectarse a la base de datos.";
-    conn.query("UPDATE clientes SET ? WHERE id = ?", [cliente_json, cliente_json.id], (err, rows) => {
-      if (err) {
-        console.log(err);
-        throw "Error en consulta de base de datos.";
-      }
-      obj_respuesta.exito = true;
-      obj_respuesta.mensaje = "Cliente Modificado!";
-      obj_respuesta.status = 200;
-      response.status(obj_respuesta.status).json(obj_respuesta);
-    });
-  });
+  }
 });
 
 // Eliminar cliente
-app.post("/eliminarCliente", (request, response) => {
+app.post("/eliminarCliente", verificar_jwt, (request, response) => {
   let obj_respuesta = {
     exito: false,
     mensaje: "No se pudo eliminar el cliente",
     status: 418,
   };
 
+  let jwt = response.jwt;
+
   let cliente_json = {};
   cliente_json.id = request.body.id;
 
-  request.getConnection((err, conn) => {
-    if (err) throw "Error al conectarse a la base de datos.";
-    conn.query("DELETE FROM clientes WHERE id = ?", [cliente_json.id], (err, rows) => {
-      if (err) {
-        console.log(err);
-        throw "Error en consulta de base de datos.";
-      }
+  if (jwt.usuario.perfil !== "administrador") {
+    obj_respuesta.mensaje = "Usuario sin permisos!!";
+    obj_respuesta.status = 401;
+    response.status(obj_respuesta.status).json(obj_respuesta);
+  } else {
+    request.getConnection((err, conn) => {
+      if (err) throw "Error al conectarse a la base de datos.";
+      conn.query("DELETE FROM clientes WHERE id = ?", [cliente_json.id], (err, rows) => {
+        if (err) {
+          console.log(err);
+          throw "Error en consulta de base de datos.";
+        }
+      });
     });
-  });
 
-  request.getConnection((err, conn) => {
-    if (err) throw "Error al conectarse a la base de datos.";
-    conn.query("DELETE FROM controles WHERE id = ?", [cliente_json.id], (err, rows) => {
-      if (err) {
-        console.log(err);
-        throw "Error en consulta de base de datos.";
-      }
-      obj_respuesta.exito = true;
-      obj_respuesta.mensaje = "Cliente Eliminado!";
-      obj_respuesta.status = 200;
-      response.status(obj_respuesta.status).json(obj_respuesta);
+    request.getConnection((err, conn) => {
+      if (err) throw "Error al conectarse a la base de datos.";
+      conn.query("DELETE FROM controles WHERE id = ?", [cliente_json.id], (err, rows) => {
+        if (err) {
+          console.log(err);
+          throw "Error en consulta de base de datos.";
+        }
+        obj_respuesta.exito = true;
+        obj_respuesta.mensaje = "Cliente Eliminado!";
+        obj_respuesta.status = 200;
+        response.status(obj_respuesta.status).json(obj_respuesta);
+      });
     });
-  });
+  }
 });
 
 // CRUD CONTROLES **************************************************************************************
 // Agregar control
-app.post("/agregarControl", (request, response) => {
+app.post("/agregarControl", verificar_jwt, (request, response) => {
   let obj_respuesta = {
     exito: false,
     mensaje: "No se pudo agregar el control",
     status: 418,
   };
+
+  let jwt = response.jwt;
 
   let control_json = {};
   control_json.id = request.body.id;
@@ -376,23 +407,29 @@ app.post("/agregarControl", (request, response) => {
   control_json.muslos = request.body.muslos;
   control_json.objetivo = request.body.objetivo;
 
-  request.getConnection((err, conn) => {
-    if (err) throw "Error al conectarse a la base de datos.";
-    conn.query("INSERT INTO controles set ?", [control_json], (err, rows) => {
-      if (err) {
-        console.log(err);
-        throw "Error en consulta de base de datos.";
-      }
-      obj_respuesta.exito = true;
-      obj_respuesta.mensaje = "Control agregado!";
-      obj_respuesta.status = 200;
-      response.status(obj_respuesta.status).json(obj_respuesta);
+  if (jwt.usuario.perfil !== "administrador") {
+    obj_respuesta.mensaje = "Usuario sin permisos!!";
+    obj_respuesta.status = 401;
+    response.status(obj_respuesta.status).json(obj_respuesta);
+  } else {
+    request.getConnection((err, conn) => {
+      if (err) throw "Error al conectarse a la base de datos.";
+      conn.query("INSERT INTO controles set ?", [control_json], (err, rows) => {
+        if (err) {
+          console.log(err);
+          throw "Error en consulta de base de datos.";
+        }
+        obj_respuesta.exito = true;
+        obj_respuesta.mensaje = "Control agregado!";
+        obj_respuesta.status = 200;
+        response.status(obj_respuesta.status).json(obj_respuesta);
+      });
     });
-  });
+  }
 });
 
 // Listar Controles
-app.get("/listarControles", (request, response) => {
+app.get("/listarControles", verificar_jwt, (request, response) => {
   let obj_respuesta = {
     exito: false,
     mensaje: "No se encontraron controles",
@@ -400,31 +437,41 @@ app.get("/listarControles", (request, response) => {
     status: 424,
   };
 
-  request.getConnection((err, conn) => {
-    if (err) throw "Error al conectarse a la base de datos.";
-    conn.query("SELECT * FROM controles", (err, rows) => {
-      if (err) throw "Error en consulta de base de datos.";
-      if (rows.length == 0) {
-        response.status(obj_respuesta.status).json(obj_respuesta);
-      } else {
-        obj_respuesta.exito = true;
-        obj_respuesta.mensaje = "Controles encontrados!";
-        obj_respuesta.dato = rows;
-        obj_respuesta.status = 200;
-        response.status(obj_respuesta.status).json(rows);
-      }
+  let jwt = response.jwt;
+
+  if (jwt.usuario.perfil !== "administrador") {
+    obj_respuesta.mensaje = "Usuario sin permisos!!";
+    obj_respuesta.status = 401;
+    response.status(obj_respuesta.status).json(obj_respuesta);
+  } else {
+    request.getConnection((err, conn) => {
+      if (err) throw "Error al conectarse a la base de datos.";
+      conn.query("SELECT * FROM controles", (err, rows) => {
+        if (err) throw "Error en consulta de base de datos.";
+        if (rows.length == 0) {
+          response.status(obj_respuesta.status).json(obj_respuesta);
+        } else {
+          obj_respuesta.exito = true;
+          obj_respuesta.mensaje = "Controles encontrados!";
+          obj_respuesta.dato = rows;
+          obj_respuesta.status = 200;
+          response.status(obj_respuesta.status).json(rows);
+        }
+      });
     });
-  });
+  }
 });
 
 // Modificar control
-app.post("/modificarControl", (request, response) => {
+app.post("/modificarControl", verificar_jwt, (request, response) => {
   let obj_respuesta = {
     exito: false,
     mensaje: "No se pudo modificar el control",
     status: 418,
   };
 
+  let jwt = response.jwt;
+
   let control_json = {};
   control_json.id = request.body.id;
   control_json.fecha = request.body.fecha;
@@ -437,33 +484,41 @@ app.post("/modificarControl", (request, response) => {
   control_json.muslos = request.body.muslos;
   control_json.objetivo = request.body.objetivo;
 
-  request.getConnection((err, conn) => {
-    if (err) throw "Error al conectarse a la base de datos.";
-    conn.query(
-      "UPDATE controles SET ? WHERE id = ? AND fecha = ?",
-      [control_json.id, control_json.fecha],
-      (err, rows) => {
-        if (err) {
-          console.log(err);
-          throw "Error en consulta de base de datos.";
+  if (jwt.usuario.perfil !== "administrador") {
+    obj_respuesta.mensaje = "Usuario sin permisos!!";
+    obj_respuesta.status = 401;
+    response.status(obj_respuesta.status).json(obj_respuesta);
+  } else {
+    request.getConnection((err, conn) => {
+      if (err) throw "Error al conectarse a la base de datos.";
+      conn.query(
+        "UPDATE controles SET ? WHERE id = ? AND fecha = ?",
+        [control_json.id, control_json.fecha],
+        (err, rows) => {
+          if (err) {
+            console.log(err);
+            throw "Error en consulta de base de datos.";
+          }
+          obj_respuesta.exito = true;
+          obj_respuesta.mensaje = "Control modificado!";
+          obj_respuesta.status = 200;
+          response.status(obj_respuesta.status).json(obj_respuesta);
         }
-        obj_respuesta.exito = true;
-        obj_respuesta.mensaje = "Control modificado!";
-        obj_respuesta.status = 200;
-        response.status(obj_respuesta.status).json(obj_respuesta);
-      }
-    );
-  });
+      );
+    });
+  }
 });
 
 // Eliminar control
-app.post("/eliminarControl", (request, response) => {
+app.post("/eliminarControl", verificar_jwt, (request, response) => {
   let obj_respuesta = {
     exito: false,
     mensaje: "No se pudo eliminar el control",
     status: 418,
   };
 
+  let jwt = response.jwt;
+
   let control_json = {};
   control_json.id = request.body.id;
   control_json.fecha = request.body.fecha;
@@ -476,26 +531,31 @@ app.post("/eliminarControl", (request, response) => {
   control_json.muslos = request.body.muslos;
   control_json.objetivo = request.body.objetivo;
 
-  request.getConnection((err, conn) => {
-    if (err) throw "Error al conectarse a la base de datos.";
-    conn.query(
-      "DELETE FROM controles WHERE id = ? AND fecha = ?",
-      [control_json.id, control_json.fecha],
-      (err, rows) => {
-        if (err) {
-          console.log(err);
-          throw "Error en consulta de base de datos.";
+  if (jwt.usuario.perfil !== "administrador") {
+    obj_respuesta.mensaje = "Usuario sin permisos!!";
+    obj_respuesta.status = 401;
+    response.status(obj_respuesta.status).json(obj_respuesta);
+  } else {
+    request.getConnection((err, conn) => {
+      if (err) throw "Error al conectarse a la base de datos.";
+      conn.query(
+        "DELETE FROM controles WHERE id = ? AND fecha = ?",
+        [control_json.id, control_json.fecha],
+        (err, rows) => {
+          if (err) {
+            console.log(err);
+            throw "Error en consulta de base de datos.";
+          }
+          obj_respuesta.exito = true;
+          obj_respuesta.mensaje = "Control eliminado!";
+          obj_respuesta.status = 200;
+          response.status(obj_respuesta.status).json(obj_respuesta);
         }
-        obj_respuesta.exito = true;
-        obj_respuesta.mensaje = "Control eliminado!";
-        obj_respuesta.status = 200;
-        response.status(obj_respuesta.status).json(obj_respuesta);
-      }
-    );
-  });
+      );
+    });
+  }
 });
 
 app.listen(app.get("puerto"), () => {
   console.log("Servidor corriendo sobre puerto:", app.get("puerto"));
 });
-//# sourceMappingURL=servidor.js.map
